Update to Neon v1.0.

+++ src/context/page.rs
@@ -354,7 +354,7 @@ impl PageSequence{
 
 pub fn pages_arg(cx: &mut FunctionContext, idx: i32, canvas:&BoxedCanvas) -> NeonResult<PageSequence> {
   let engine = canvas.borrow().engine;
-  let pages = cx.argument::<JsArray>(idx)?
+  let pages = cx.argument::<JsArray>(idx as usize)?
       .to_vec(cx)?
       .iter()
       .map(|obj| obj.downcast::<BoxedContext2D, _>(cx))
